backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn==1.6.1
xgboost>=2.0.0
python-multipart==0.0.6
aiofiles==23.2.1
pillow>=10.0.0
scipy>=1.10.0
tensorflow>=2.13.0
torch>=2.0.0
torchvision>=0.15.0
sqlalchemy>=2.0.0
alembic==1.13.0
python-dotenv==1.0.0
joblib>=1.3.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
bcrypt>=4.0.0
psycopg2-binary>=2.9.0

